+++ telekit/example/example_handlers/entry.py
@@ -31,6 +31,7 @@ class UserData:
     def set_age(self, value: int):
         self.ages[self.chat_id] = value
     
+import gc
 
 class EntryHandler(telekit.Handler):
 
@@ -49,6 +50,16 @@ class EntryHandler(telekit.Handler):
 
     def handle(self) -> None:
         self._user_data = UserData(self.message.chat.id)
+        @self.chain.on_timeout(5)
+        def _on_timeout():
+            print(gc.collect())
+            for o in gc.get_objects():
+                if isinstance(o, (
+                    telekit.Handler,
+                    telekit.Chain,
+                    telekit.Handler
+                )):
+                    print(o)
         self.entry_name()
 
     # -------------------------------